Clean up.
git-svn-id: http://es-operating-system.googlecode.com/svn/trunk/esidl@412 d15c3b8f-5b43-0410-be93-4d5a55ce2954

+++ include/string.h
@@ -27,7 +27,7 @@
 class StringOffsetter : public Visitor
 {
     std::map<std::string, size_t>& dict;
-    u32 offset;
+    size_t offset;
 
 public:
     StringOffsetter(std::map<std::string, size_t>& dict, size_t base) :
@@ -36,7 +36,7 @@ public:
     {
     }
 
-    u32 getOffset() const
+    size_t getOffset() const
     {
         return offset;
     }
